+++ lib/Models/Terria.ts
@@ -673,13 +673,6 @@ export default class Terria {
     }
   }
 
-  setBaseMaps(baseMaps: BaseMapModel[]): void {
-    processBaseMaps(baseMaps, this);
-    if (!this.mainViewer.baseMap) {
-      this.loadPersistedOrInitBaseMap();
-    }
-  }
-
   /**
    * Loading of basemap is done in the following order
    * - first try to load basemap set in the URL parameter `baseMapId`
@@ -796,6 +789,10 @@ export default class Terria {
 
     if (this.baseMaps.length === 0) {
       processBaseMaps(defaultBaseMaps(this), this);
+    }
+
+    if (!this.mainViewer.baseMap) {
+      this.loadPersistedOrInitBaseMap();
     }
   }
 
@@ -973,7 +970,7 @@ export default class Terria {
       Array.isArray(initData.baseMaps) &&
       initData.baseMaps.length > 0
     ) {
-      this.setBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps));
+      processBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps), this);
     }
 
     if (isJsonObject(initData.homeCamera)) {