+++ lib/Models/Terria.ts
@@ -673,30 +673,44 @@ export default class Terria {
     }
   }
 
-  setBaseMaps(baseMaps: BaseMapModel[]): void {
-    processBaseMaps(baseMaps, this);
-    if (!this.mainViewer.baseMap) {
-      this.loadPersistedOrInitBaseMap();
-    }
-  }
-
+  /**
+   * Loading of basemap is done in the following order
+   * - first try to load basemap set in the URL parameter `baseMapId`
+   * - if it is not defined try to load persisted basemap
+   * - lastly try to load init basemap
+   */
   @action
   loadPersistedOrInitBaseMap(): void {
-    const persistedBaseMapId = this.getLocalProperty("basemap");
-    const baseMapSearch = this.baseMaps.find(
-      baseMap => baseMap.mappable.uniqueId === persistedBaseMapId
-    );
-    if (baseMapSearch) {
-      this.mainViewer.baseMap = baseMapSearch.mappable;
-    } else {
-      console.error(
-        `Couldn't find a basemap for unique id ${persistedBaseMapId}. Trying to load init base map.`
-      );
+    const baseMapId = this.userProperties.get("baseMapId");
+    if (baseMapId) {
       const baseMapSearch = this.baseMaps.find(
-        baseMap => baseMap.mappable.uniqueId === this.initBaseMapId
+        baseMap => baseMap.mappable.uniqueId === baseMapId
       );
       if (baseMapSearch) {
         this.mainViewer.baseMap = baseMapSearch.mappable;
+      } else {
+        console.error(
+          `Couldn't find a basemap for unique id ${baseMapId}. Trying to load persisted base map.`
+        );
+      }
+    }
+    if (!isDefined(this.mainViewer.baseMap)) {
+      const persistedBaseMapId = this.getLocalProperty("basemap");
+      const baseMapSearch = this.baseMaps.find(
+        baseMap => baseMap.mappable.uniqueId === persistedBaseMapId
+      );
+      if (baseMapSearch) {
+        this.mainViewer.baseMap = baseMapSearch.mappable;
+      } else {
+        console.error(
+          `Couldn't find a basemap for unique id ${persistedBaseMapId}. Trying to load init base map.`
+        );
+        const baseMapSearch = this.baseMaps.find(
+          baseMap => baseMap.mappable.uniqueId === this.initBaseMapId
+        );
+        if (baseMapSearch) {
+          this.mainViewer.baseMap = baseMapSearch.mappable;
+        }
       }
     }
   }
@@ -775,6 +789,10 @@ export default class Terria {
 
     if (this.baseMaps.length === 0) {
       processBaseMaps(defaultBaseMaps(this), this);
+    }
+
+    if (!this.mainViewer.baseMap) {
+      this.loadPersistedOrInitBaseMap();
     }
   }
 
@@ -952,7 +970,7 @@ export default class Terria {
       Array.isArray(initData.baseMaps) &&
       initData.baseMaps.length > 0
     ) {
-      this.setBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps));
+      processBaseMaps(<BaseMapModel[]>(<unknown>initData.baseMaps), this);
     }
 
     if (isJsonObject(initData.homeCamera)) {