+++ CHANGES.md
@@ -4,6 +4,11 @@ Change Log
 ### MobX Development
 
 #### next release (8.0.0-alpha.69)
+
+- **Breaking changes**: 
+  - Basemaps are now configured through catalog JSON instead of TerriaMap - see https://github.com/TerriaJS/terriajs/blob/next/doc/customizing/initialization-files.md#base-maps 
+
+* Make base maps configurable from init files and update documentation for init files [#5140](https://github.com/TerriaJS/terriajs/pull/5140).
 * [The next improvement]
 
 #### 8.0.0-alpha.68