+++ CHANGES.md
@@ -4,6 +4,9 @@ Change Log
 ### MobX Development
 
 #### next release (8.0.0-alpha.70)
+* Fix load of persisted basemap
+* Fix sharing of base map
+* Add `baseMapId` URL parameter to allow setting of basemap
 * [The next improvement]
 
 #### 8.0.0-alpha.69